fig.update_layout(width=1500, height=2000) This line of code modifies the layout of the figure created by make_subplots() by setting its width and height. Here’s a detailed breakdown of what’s happening:
1. fig.update_layout(...)
fig: This refers to the figure object created by make_subplots(). It contains the data and layout information for all the subplots.

update_layout(): This is a method used in Plotly to modify the layout of the figure. Layout in Plotly refers to non-data aspects of the plot, such as the size, title, axis labels, background color, etc.

2. width=1500, height=2000
width=1500: This sets the width of the entire figure to 1500 pixels. It determines how wide the plot will appear when rendered on the screen.

height=2000: This sets the height of the entire figure to 2000 pixels. It determines how tall the plot will appear when rendered on the screen.

Purpose of update_layout() in This Context:
Customization: The default size of the plot may not be ideal for displaying the data clearly. By updating the layout, the figure’s size is explicitly defined, ensuring that the subplots are large enough to be easily readable and properly spaced.

Adjusting Aspect Ratio: Given that there are 8 vertical subplots, increasing the height allows more space for each subplot, making the plot less cramped and more legible.

Better Visibility: By adjusting the width and height, the plot is customized to fit the content, ensuring that axis labels, titles, and data points don't overlap or get cut off.

In Summary:
This line of code ensures that the figure is displayed with a width of 1500 pixels and a height of 2000 pixels, which provides enough space for each of the 8 subplots to be clearly visible without overlapping or being too small to interpret effectively.